Rev 0.4:
        Encoder and Dome Threads running.
        Shell running out of thread (used the loop function as idle thread).
        Added Encoder simulation via timer for debugging. This feature is activate via switch.
        Added Semaphore to count # of turning steps.
Added code to return angle, but float not working.
Added code to configure software, but issue on int/float conversion.
        TO DO:
        1) Check the move left /right of given #of steps; -> Not working ON HOLD!
        2) Return angle and not counts; -> Buggy
        3) Move all encoder configuration defines on variable to allow re/configuration of the code;
        4) Allow code configuration; -> Buggy

+++ dome.cpp
@@ -161,8 +161,10 @@ NIL_THREAD(DomeThread, arg)
 		avrPrintf("Using Dome Thread\nReceived data is: ");
 		avrPrintf(*p);
 		avrPrintf(CR);
-		unsigned long posInitial = Encoder.Position();
-		unsigned long posFinal;
+		//unsigned long posInitial = Encoder.Position();
+		//unsigned long posFinal
+		uint32_t posInitial = Encoder.Position();
+		uint32_t posFinal;
 		avrPrintf("Dome Initial Position: ");
 		avrPrintf(posInitial);
 		avrPrintf(CR);
